+++ shardingsphere-features/shardingsphere-encrypt/shardingsphere-encrypt-common/src/test/resources/META-INF/services/org.apache.shardingsphere.encrypt.spi.EncryptAlgorithm
@@ -15,5 +15,5 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.encrypt.strategy.fixture.TestEncryptAlgorithm
-org.apache.shardingsphere.encrypt.strategy.fixture.TestQueryAssistedEncryptAlgorithm
+org.apache.shardingsphere.encrypt.algorithm.fixture.TestEncryptAlgorithm
+org.apache.shardingsphere.encrypt.algorithm.fixture.TestQueryAssistedEncryptAlgorithm
